Move clustering end-to-end test into separate file
* Moved testValuesRemainClusteredAfterTraining() from cluster_wrapper_test.py
  to cluster_integration_test.py
* Updated BUILD with new cluster_integration_test py_test target
